Generalize command execution method and extract it into the CommandExecutor class
