+++ src/app/ngx-editor/ngx-editor.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, HostListener, Input, Output, ElementRef, EventEmitter, ViewChild } from '@angular/core';
 import { ngxEditorConfig } from './ngx-editor.defaults';
+import {CommandExecutor} from './command-executor';
 
 @Component({
   selector: 'app-ngx-editor',
@@ -18,6 +19,7 @@ export class NgxEditorComponent implements OnInit {
   _resizer: string;
   ngxMessage: string;
   enableToolbar = false;
+  commandExecutor = new CommandExecutor(document);
 
   @Input() editable: boolean;
   @Input() spellcheck: boolean;
@@ -89,50 +91,8 @@ export class NgxEditorComponent implements OnInit {
    * editor actions
    */
   executeCommand(commandName) {
-    document.execCommand(commandName, false, null);
+    this.commandExecutor.execute(commandName);
   }
-
-  // blockquote
-  blockQuote() {
-    document.execCommand('formatBlock', false, '<blockquote>');
-  }
-
-  removeQuote() {
-    document.execCommand('formatBlock', false, 'div');
-  }
-
-  // insert link
-  createLink() {
-    const selection = document.getSelection();
-
-    if (selection.anchorNode.parentElement.tagName === 'A') {
-      const linkURL = prompt('Enter URL', selection.anchorNode.parentElement.getAttribute('href'));
-      if (linkURL) {
-        document.execCommand('createLink', false, linkURL);
-      }
-    } else {
-      if (selection['type'] === 'None') {
-        this.createMessage('No selection made');
-      } else {
-        const linkURL = prompt('Enter URL', 'http://');
-        if (linkURL) {
-          document.execCommand('createLink', false, linkURL);
-        }
-      }
-    }
-  }
-
-  // insert image
-  insertImage() {
-    const imageURI = prompt('Enter Image URL', 'http://');
-    if (imageURI) {
-      const inserted = document.execCommand('insertImage', false, imageURI);
-      if (!inserted) {
-        this.createMessage('Invalid URL');
-      }
-    }
-  }
-
   /*
    * message box
    */
